parser compiler bug fixes, independent routing and ord1 version commit
